requirements.txt
# Core ML libraries
torch>=2.0.0
transformer-lens>=1.17.0
sae-lens>=3.0.0
transformers>=4.30.0

# Data and computation
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
datasets>=2.14.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0

# Optional: For better performance
accelerate>=0.20.0
einops>=0.6.0
